+++ CHANGELOG.md
@@ -82,6 +82,7 @@
 - PIM-9925: Fix roles that couldn't contain dashes in their codes 
 - PIM-9933: Fix delete category menu that stays displayed on screen
 - PIM-9949: Fix category edit page to use catalog locale
+- PIM-9948: Fix performance issue on product model import
 
 ## New features